# my-page-api/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: JPro My Page API
  description: API for JPro My Page
  version: 1.0.0
paths:
...
  /cabin-lottery/drawings/{drawingId}:
    get:
      tags:
        - CabinLottery
      operationId: getDrawing
      summary: Get drawing details
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Drawing details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinDrawing'
  /cabin-lottery/drawings/{drawingId}/periods:
    get:
      tags:
        - CabinLottery
      operationId: getDrawingPeriods
      summary: Get all periods for a drawing
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of periods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CabinPeriod'
  /cabin-lottery/drawings/{drawingId}/wishes:
    post:
      tags:
        - CabinLottery
      operationId: submitWishes
      summary: Submit wishes for a drawing
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkCreateWishes'
      responses:
        '200':
          description: Wishes created
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CabinWish'
  /cabin-lottery/drawings/{drawingId}/my-wishes:
    get:
      tags:
        - CabinLottery
      operationId: getMyWishes
      summary: Get current user's wishes
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of wishes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CabinWish'
  /cabin-lottery/drawings/{drawingId}/my-allocations:
    get:
      tags:
        - CabinLottery
      operationId: getMyAllocations
      summary: Get current user's allocations
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CabinAllocation'
  /cabin-lottery/drawings/{drawingId}/allocations:
    get:
      tags:
        - CabinLottery
      operationId: getDrawingAllocations
      summary: Get all allocations for a drawing
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: executionId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CabinAllocation'
...
  /cabin-lottery/admin/drawings:
    post:
      tags:
        - CabinLottery
      operationId: createAdminDrawing
      summary: Create a new cabin drawing (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                season:
                  type: string
              required:
                - season
      responses:
        '200':
          description: Drawing created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinDrawing'
        '400':
          description: Bad request
    get:
      tags:
        - CabinLottery
      operationId: getAllDrawings
      summary: Get all drawings (admin only)
      responses:
        '200':
          description: List of all drawings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CabinDrawing'
  /cabin-lottery/admin/drawings/{drawingId}:
    get:
      tags:
        - CabinLottery
      operationId: getAdminDrawing
      summary: Get drawing by ID (admin only)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Drawing details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinDrawing'
        '404':
          description: Drawing not found
    delete:
      tags:
        - CabinLottery
      operationId: deleteDrawing
      summary: Delete a drawing (admin only, only if not drawn or published)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Drawing deleted
        '400':
          description: Bad request
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/periods:
    post:
      tags:
        - CabinLottery
      operationId: createPeriod
      summary: Add a period to a drawing (admin only)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePeriod'
      responses:
        '200':
          description: Period created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinPeriod'
        '400':
          description: Bad request
        '404':
          description: Drawing not found
    get:
      tags:
        - CabinLottery
      operationId: getAdminPeriods
      summary: Get all periods for a drawing (admin only)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of periods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CabinPeriod'
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/periods/{periodId}:
    put:
      tags:
        - CabinLottery
      operationId: updatePeriod
      summary: Update a period (admin only)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: periodId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePeriod'
      responses:
        '200':
          description: Period updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinPeriod'
        '400':
          description: Bad request
        '404':
          description: Period not found
    delete:
      tags:
        - CabinLottery
      operationId: deletePeriod
      summary: Delete a period (admin only, only if drawing is OPEN)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: periodId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Period deleted
        '400':
          description: Bad request
        '404':
          description: Period not found
  /cabin-lottery/admin/drawings/{drawingId}/periods/bulk:
    post:
      tags:
        - CabinLottery
      operationId: bulkCreatePeriods
      summary: Bulk create periods (admin only, Wednesday to Wednesday)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                startDate:
                  type: string
                  format: date
                endDate:
                  type: string
                  format: date
                descriptionTemplate:
                  type: string
              required:
                - startDate
                - endDate
      responses:
        '200':
          description: Periods created
          content:
            application/json:
              schema:
                type: object
                properties:
                  periodsCreated:
                    type: integer
                  periods:
                    type: array
                    items:
                      $ref: '#/components/schemas/CabinPeriod'
                required:
                  - periodsCreated
                  - periods
        '400':
          description: Bad request
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/open:
    post:
      tags:
        - CabinLottery
      operationId: openDrawing
      summary: Open a drawing (admin only, DRAFT -> OPEN, makes it visible to users)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Drawing opened
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinDrawing'
        '400':
          description: Bad request
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/revert-to-draft:
    post:
      tags:
        - CabinLottery
      operationId: revertToDraft
      summary: Revert drawing to DRAFT (admin only, OPEN/LOCKED -> DRAFT)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Drawing reverted to draft
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinDrawing'
        '400':
          description: Bad request
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/revert-to-locked:
    post:
      tags:
        - CabinLottery
      operationId: revertToLocked
      summary: Revert drawing to LOCKED (admin only, DRAWN -> LOCKED)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Drawing reverted to locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinDrawing'
        '400':
          description: Bad request
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/lock:
    post:
      tags:
        - CabinLottery
      operationId: lockDrawing
      summary: Lock a drawing (admin only, no more wish changes allowed)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Drawing locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinDrawing'
        '400':
          description: Bad request
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/unlock:
    post:
      tags:
        - CabinLottery
      operationId: unlockDrawing
      summary: Unlock a drawing (admin only, allow wish changes again)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Drawing unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinDrawing'
        '400':
          description: Bad request
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/draw:
    post:
      tags:
        - CabinLottery
      operationId: performDrawing
      summary: Perform the snake draft lottery (admin only)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: seed
          in: query
          required: false
          schema:
            type: integer
            format: int64
          description: Optional random seed for reproducibility
      responses:
        '200':
          description: Drawing performed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DrawingResult'
        '400':
          description: Bad request
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/publish:
    post:
      tags:
        - CabinLottery
      operationId: publishDrawing
      summary: Publish drawing results to users (admin only)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: executionId
          in: query
          required: true
          schema:
            type: string
            format: uuid
          description: The execution ID to publish
      responses:
        '200':
          description: Drawing published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CabinDrawing'
        '400':
          description: Bad request
        '404':
          description: Drawing or execution not found
  /cabin-lottery/admin/drawings/{drawingId}/executions/{executionId}:
    delete:
      tags:
        - CabinLottery
      operationId: deleteExecution
      summary: Delete an execution (admin only, only if drawing not published)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: executionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Execution deleted
        '400':
          description: Bad request
        '404':
          description: Execution not found
  /cabin-lottery/admin/drawings/{drawingId}/wishes:
    get:
      tags:
        - CabinLottery
      operationId: getAllWishes
      summary: Get all wishes for a drawing (admin only)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of all wishes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CabinWish'
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/allocations:
    get:
      tags:
        - CabinLottery
      operationId: getAdminAllocations
      summary: Get allocations for a drawing (admin only, published or latest execution)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: executionId
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Optional execution ID, defaults to published or latest
      responses:
        '200':
          description: List of allocations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CabinAllocation'
        '404':
          description: Drawing not found
  /cabin-lottery/admin/drawings/{drawingId}/import:
    post:
      tags:
        - CabinLottery
      operationId: importWishes
      summary: Import wishes from CSV file (admin only, Google Forms format)
      parameters:
        - name: drawingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '200':
          description: Wishes imported
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalLines:
                    type: integer
                  successCount:
                    type: integer
                  errorCount:
                    type: integer
                  errors:
                    type: array
                    items:
                      type: string
                required:
                  - totalLines
                  - successCount
                  - errorCount
                  - errors
        '400':
          description: Bad request
        '404':
          description: Drawing not found
...
components:
  schemas:
...
    CabinDrawing:
      type: object
      properties:
        id:
          type: string
          format: uuid
        season:
          type: string
        status:
          type: string
          enum: [DRAFT, OPEN, LOCKED, DRAWN, PUBLISHED]
        createdAt:
          type: string
          format: date-time
        lockedAt:
          type: string
          format: date-time
        publishedAt:
          type: string
          format: date-time
        publishedExecutionId:
          type: string
          format: uuid
        publishedBy:
          type: integer
          format: int64
        publishedByName:
          type: string
        periods:
          type: array
          items:
            $ref: '#/components/schemas/CabinPeriod'
        executions:
          type: array
          items:
            $ref: '#/components/schemas/CabinDrawingExecution'
        bookingWarnings:
          type: array
          items:
            type: string
      required:
        - id
        - season
        - status
        - createdAt
    CabinPeriod:
      type: object
      properties:
        id:
          type: string
          format: uuid
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
        comment:
          type: string
        sortOrder:
          type: integer
      required:
        - id
        - startDate
        - endDate
        - description
        - sortOrder
    CreatePeriod:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
        comment:
          type: string
        sortOrder:
          type: integer
          default: 0
      required:
        - startDate
        - endDate
        - description
    CabinWish:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: integer
          format: int64
        userName:
          type: string
        userEmail:
          type: string
        periodId:
          type: string
          format: uuid
        periodDescription:
          type: string
        priority:
          type: integer
        desiredApartmentIds:
          type: array
          items:
            type: integer
            format: int64
        desiredApartmentNames:
          type: array
          items:
            type: string
        comment:
          type: string
      required:
        - id
        - userId
        - userName
        - userEmail
        - periodId
        - periodDescription
        - priority
        - desiredApartmentIds
        - desiredApartmentNames
...
    BulkCreateWishes:
      type: object
      properties:
        wishes:
          type: array
          items:
            $ref: '#/components/schemas/CreateCabinWish'
      required:
        - wishes
    CabinAllocation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        periodId:
          type: string
          format: uuid
        periodDescription:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        apartmentId:
          type: integer
          format: int64
        apartmentName:
          type: string
        apartmentSortOrder:
          type: integer
        userId:
          type: integer
          format: int64
        userName:
          type: string
        userEmail:
          type: string
        allocationType:
          type: string
        comment:
          type: string
        allocatedAt:
          type: string
          format: date-time
      required:
        - id
        - periodId
        - periodDescription
        - startDate
        - endDate
        - apartmentId
        - apartmentName
        - userId
        - userName
        - userEmail
        - allocationType
        - allocatedAt
...
    DrawingResult:
      type: object
      properties:
        drawingId:
          type: string
          format: uuid
        executionId:
          type: string
          format: uuid
        season:
          type: string
        drawnAt:
          type: string
          format: date-time
        allocations:
          type: array
          items:
            $ref: '#/components/schemas/CabinAllocation'
        statistics:
          $ref: '#/components/schemas/DrawingStatistics'
        auditLog:
          type: array
          items:
            type: string
      required:
        - drawingId
        - executionId
        - season
        - drawnAt
        - allocations
        - statistics
        - auditLog
...